Fix type inference of createSlice
Use a lookup type to infer the type of the keyed property in createSlice. Currently, type inference for creating a new slice is broken - if you create a slice on a valid property, it will always return {} or the type of your initialState. 

I think this was missed because in your unit tests `test_slicing.ts` you declare the type of the sliced store explicitly, so that you can re-use variables between your tests with beforeEach. If you set up a test completely local then the bug is visible.